filter returns accommodations with the rating

+++ src/main/java/com/booking/domain/AccommodaitonReview.java
@@ -6,7 +6,7 @@ import java.util.Date;
 public class AccommodaitonReview extends Review{
     private long AccommodationId;
 
-    public AccommodaitonReview(Long reviewId, LocalDate reviewDate, String description, float rating, long accommodationId) {
+    public AccommodaitonReview(Long reviewId, LocalDate reviewDate, String description, Double rating, long accommodationId) {
         super(reviewId, reviewDate, description, rating);
         AccommodationId = accommodationId;
     }
