+++ src/main/java/com/booking/domain/AccommodationReview.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public class AccommodationReview extends Review{
     private Long AccommodationId;
 
-    public AccommodationReview(Long reviewId, LocalDate reviewDate, String description, float rating, long accommodationId) {
+    public AccommodationReview(Long reviewId, LocalDate reviewDate, String description, Double rating, long accommodationId) {
         super(reviewId, reviewDate, description, rating);
         AccommodationId = accommodationId;
     }
